사원 초대 메시지

+++ J2SM/src/main/java/kr/co/J2SM/dto/company/CompanyDTO.java
@@ -9,6 +9,6 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class CompanyDTO {
-    private Long cno;
+    private int cno;
     private String companyName;
 }
